修改包头编码解码方式，活动落实

+++ src/main/java/cn/nukkit/network/protocol/AddEntityPacket.java
@@ -9,10 +9,10 @@ import cn.nukkit.utils.Binary;
  * Nukkit Project
  */
 public class AddEntityPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.ADD_ENTITY_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.ADD_ENTITY_PACKET;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
