+++ src/main/java/cn/nukkit/network/protocol/AddPaintingPacket.java
@@ -5,7 +5,7 @@ package cn.nukkit.network.protocol;
  */
 public class AddPaintingPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.ADD_PAINTING_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.ADD_PAINTING_PACKET;
 
     public long entityUniqueId;
     public long entityRuntimeId;
@@ -31,7 +31,7 @@ public class AddPaintingPacket extends DataPacket {
     }
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
