+++ src/main/java/cn/nukkit/network/protocol/AddPlayerPacket.java
@@ -11,10 +11,10 @@ import java.util.UUID;
  * Nukkit Project
  */
 public class AddPlayerPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.ADD_PLAYER_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.ADD_PLAYER_PACKET;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
