+++ src/main/java/cn/nukkit/network/protocol/AnimatePacket.java
@@ -5,7 +5,7 @@ package cn.nukkit.network.protocol;
  */
 public class AnimatePacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.ANIMATE_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.ANIMATE_PACKET;
 
     public long eid;
     public int action;
@@ -31,7 +31,7 @@ public class AnimatePacket extends DataPacket {
     }
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
