+++ src/main/java/cn/nukkit/network/protocol/BlockEntityDataPacket.java
@@ -7,7 +7,7 @@ import cn.nukkit.math.BlockVector3;
  * Nukkit Project
  */
 public class BlockEntityDataPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.BLOCK_ENTITY_DATA_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.BLOCK_ENTITY_DATA_PACKET;
 
     public int x;
     public int y;
@@ -15,7 +15,7 @@ public class BlockEntityDataPacket extends DataPacket {
     public byte[] namedTag;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
