+++ src/main/java/cn/nukkit/network/protocol/BlockPickRequestPacket.java
@@ -4,7 +4,7 @@ import cn.nukkit.math.BlockVector3;
 
 public class BlockPickRequestPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.BLOCK_PICK_REQUEST_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.BLOCK_PICK_REQUEST_PACKET;
 
     public int x;
     public int y;
@@ -13,7 +13,7 @@ public class BlockPickRequestPacket extends DataPacket {
     public int selectedSlot;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
