+++ src/main/java/cn/nukkit/network/protocol/CommandRequestPacket.java
@@ -10,7 +10,7 @@ import java.util.UUID;
  */
 public class CommandRequestPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.COMMAND_REQUEST_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.COMMAND_REQUEST_PACKET;
 
     public static final int TYPE_PLAYER = 0;
     public static final int TYPE_COMMAND_BLOCK = 1;
@@ -28,7 +28,7 @@ public class CommandRequestPacket extends DataPacket {
     public CommandOriginData data;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
