+++ src/main/java/cn/nukkit/network/protocol/ContainerClosePacket.java
@@ -5,10 +5,10 @@ package cn.nukkit.network.protocol;
  * Nukkit Project
  */
 public class ContainerClosePacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.CONTAINER_CLOSE_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.CONTAINER_CLOSE_PACKET;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
