+++ src/main/java/cn/nukkit/network/protocol/ContainerOpenPacket.java
@@ -7,10 +7,10 @@ import cn.nukkit.math.BlockVector3;
  * Nukkit Project
  */
 public class ContainerOpenPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.CONTAINER_OPEN_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.CONTAINER_OPEN_PACKET;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
