+++ src/main/java/cn/nukkit/network/protocol/EntityFallPacket.java
@@ -1,7 +1,7 @@
 package cn.nukkit.network.protocol;
 
 public class EntityFallPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.ENTITY_FALL_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.ENTITY_FALL_PACKET;
 
     public long eid;
     public float fallDistance;
@@ -20,7 +20,7 @@ public class EntityFallPacket extends DataPacket {
     }
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
 }
