+++ src/main/java/cn/nukkit/network/protocol/EntityPickRequestPacket.java
@@ -2,10 +2,10 @@ package cn.nukkit.network.protocol;
 
 public class EntityPickRequestPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.ENTITY_PICK_REQUEST_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.ENTITY_PICK_REQUEST_PACKET;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
