+++ src/main/java/cn/nukkit/network/protocol/FullChunkDataPacket.java
@@ -5,10 +5,10 @@ package cn.nukkit.network.protocol;
  * Nukkit Project
  */
 public class FullChunkDataPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.FULL_CHUNK_DATA_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.FULL_CHUNK_DATA_PACKET;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
