+++ src/main/java/cn/nukkit/network/protocol/HurtArmorPacket.java
@@ -5,7 +5,7 @@ package cn.nukkit.network.protocol;
  */
 public class HurtArmorPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.HURT_ARMOR_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.HURT_ARMOR_PACKET;
 
     public int health;
 
@@ -21,7 +21,7 @@ public class HurtArmorPacket extends DataPacket {
     }
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
