+++ src/main/java/cn/nukkit/network/protocol/InteractPacket.java
@@ -5,7 +5,7 @@ package cn.nukkit.network.protocol;
  */
 public class InteractPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.INTERACT_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.INTERACT_PACKET;
 
     public static final int ACTION_VEHICLE_EXIT = 3;
     public static final int ACTION_MOUSEOVER = 4;
@@ -29,7 +29,7 @@ public class InteractPacket extends DataPacket {
     }
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
