+++ src/main/java/cn/nukkit/network/protocol/InventoryContentPacket.java
@@ -7,10 +7,10 @@ import cn.nukkit.item.Item;
  * Nukkit Project
  */
 public class InventoryContentPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.INVENTORY_CONTENT_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.INVENTORY_CONTENT_PACKET;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
