+++ src/main/java/cn/nukkit/network/protocol/ItemFrameDropItemPacket.java
@@ -7,7 +7,7 @@ import cn.nukkit.math.BlockVector3;
  */
 public class ItemFrameDropItemPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.ITEM_FRAME_DROP_ITEM_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.ITEM_FRAME_DROP_ITEM_PACKET;
 
     public int x;
     public int y;
@@ -27,7 +27,7 @@ public class ItemFrameDropItemPacket extends DataPacket {
     }
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
 }
