+++ src/main/java/cn/nukkit/network/protocol/LevelEventPacket.java
@@ -7,7 +7,7 @@ import cn.nukkit.math.Vector3f;
  * Nukkit Project
  */
 public class LevelEventPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.LEVEL_EVENT_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.LEVEL_EVENT_PACKET;
 
     public static final int EVENT_SOUND_CLICK = 1000;
     public static final int EVENT_SOUND_CLICK_FAIL = 1001;
@@ -98,7 +98,7 @@ public class LevelEventPacket extends DataPacket {
     public int data = 0;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
