+++ src/main/java/cn/nukkit/network/protocol/MobArmorEquipmentPacket.java
@@ -7,10 +7,10 @@ import cn.nukkit.item.Item;
  * Nukkit Project
  */
 public class MobArmorEquipmentPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.MOB_ARMOR_EQUIPMENT_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.MOB_ARMOR_EQUIPMENT_PACKET;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
