+++ src/main/java/cn/nukkit/network/protocol/MobEffectPacket.java
@@ -6,10 +6,10 @@ package cn.nukkit.network.protocol;
  */
 public class MobEffectPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.MOB_EFFECT_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.MOB_EFFECT_PACKET;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
