+++ src/main/java/cn/nukkit/network/protocol/MovePlayerPacket.java
@@ -7,7 +7,7 @@ import cn.nukkit.math.Vector3f;
  */
 public class MovePlayerPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.MOVE_PLAYER_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.MOVE_PLAYER_PACKET;
 
     public static final int MODE_NORMAL = 0;
     public static final int MODE_RESET = 1;
@@ -64,7 +64,7 @@ public class MovePlayerPacket extends DataPacket {
     }
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
