+++ src/main/java/cn/nukkit/network/protocol/PlaySoundPacket.java
@@ -2,7 +2,7 @@ package cn.nukkit.network.protocol;
 
 public class PlaySoundPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.PLAY_SOUND_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.PLAY_SOUND_PACKET;
 
     public String name;
     public int x;
@@ -12,7 +12,7 @@ public class PlaySoundPacket extends DataPacket {
     public float pitch;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
