+++ src/main/java/cn/nukkit/network/protocol/PlayStatusPacket.java
@@ -5,10 +5,10 @@ package cn.nukkit.network.protocol;
  */
 public class PlayStatusPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.PLAY_STATUS_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.PLAY_STATUS_PACKET;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
