+++ src/main/java/cn/nukkit/network/protocol/PlayerInputPacket.java
@@ -5,7 +5,7 @@ package cn.nukkit.network.protocol;
  */
 public class PlayerInputPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.PLAYER_INPUT_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.PLAYER_INPUT_PACKET;
 
     public float motionX;
     public float motionY;
@@ -27,7 +27,7 @@ public class PlayerInputPacket extends DataPacket {
     }
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
