+++ src/main/java/cn/nukkit/network/protocol/PlayerListPacket.java
@@ -9,7 +9,7 @@ import java.util.UUID;
  */
 public class PlayerListPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.PLAYER_LIST_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.PLAYER_LIST_PACKET;
 
     public static final byte TYPE_ADD = 0;
     public static final byte TYPE_REMOVE = 1;
@@ -42,7 +42,7 @@ public class PlayerListPacket extends DataPacket {
     }
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
