+++ src/main/java/cn/nukkit/network/protocol/RemoveEntityPacket.java
@@ -5,12 +5,12 @@ package cn.nukkit.network.protocol;
  * Nukkit Project
  */
 public class RemoveEntityPacket extends DataPacket {
-    public static final byte NETWORK_ID = ProtocolInfo.REMOVE_ENTITY_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.REMOVE_ENTITY_PACKET;
 
     public long eid;
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
