+++ src/main/java/cn/nukkit/network/protocol/RequestChunkRadiusPacket.java
@@ -6,7 +6,7 @@ package cn.nukkit.network.protocol;
  */
 public class RequestChunkRadiusPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.REQUEST_CHUNK_RADIUS_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.REQUEST_CHUNK_RADIUS_PACKET;
 
     public int radius;
 
@@ -21,7 +21,7 @@ public class RequestChunkRadiusPacket extends DataPacket {
     }
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
