+++ src/main/java/cn/nukkit/network/protocol/ResourcePackChunkRequestPacket.java
@@ -2,7 +2,7 @@ package cn.nukkit.network.protocol;
 
 public class ResourcePackChunkRequestPacket extends DataPacket {
 
-    public static final byte NETWORK_ID = ProtocolInfo.RESOURCE_PACK_CHUNK_REQUEST_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.RESOURCE_PACK_CHUNK_REQUEST_PACKET;
 
     public String packId;
     public int chunkIndex;
@@ -21,7 +21,7 @@ public class ResourcePackChunkRequestPacket extends DataPacket {
     }
 
     @Override
-    public byte pid() {
+    public int pid() {
         return NETWORK_ID;
     }
 }
